api/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-jose==3.3.0
passlib==1.7.4
psycopg2-binary==2.9.6
pymongo==4.3.3
cassandra-driver==3.25.0
kafka-python==2.0.2
python-multipart==0.0.6
boto3==1.26.129
PyJWT==2.6.0  # Add this line for JWT support